added fetch coordinates to homepage.js and took out of map.js
